Link libcuda.so before libregent.so
When building with CUDA on a head node without a GPU, libcuda.so will not be
found in /usr/lib. Libregent.so will want to link against it, and only searches
in /usr/lib, causing linking to fail. The head node, however, does contain a
stub libcuda.so in $CUDA_HOME/lib64/stubs, so we can preemptively link with
that, and this will appease libregent.so. At runtime, on a node with a GPU, the
real libraries will be at the appropriate place, and the dynamic linker will
use those.